+++ instrumentation/logback/logback-appender-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/logback/appender/v1_0/LogbackSingletons.java
@@ -7,9 +7,9 @@ package io.opentelemetry.javaagent.instrumentation.logback.appender.v1_0;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.incubator.config.DeclarativeConfigProperties;
-import io.opentelemetry.instrumentation.api.config.IncludeExclude;
 import io.opentelemetry.instrumentation.api.incubator.config.internal.DeclarativeConfigUtil;
 import io.opentelemetry.instrumentation.logback.appender.v1_0.internal.LoggingEventMapper;
+import java.util.function.Predicate;
 
 public class LogbackSingletons {
 
@@ -35,7 +35,7 @@ public class LogbackSingletons {
         config.getBoolean("capture_logstash_marker_attributes/development", false);
     boolean captureLogstashStructuredArguments =
         config.getBoolean("capture_logstash_structured_arguments/development", false);
-    IncludeExclude mdcAttributes =
+    Predicate<String> mdcAttributes =
         LogbackConfig.create(GlobalOpenTelemetry.get()).getMdcAttributes();
 
     mapper =